Improve setup (#22)
* Add note about recommended project name length

Thanks to this, resources on AWS have shorter names and are visible in
the e.g. CodePipeline.

* Remove files idempotently

Otherwise, running `setup.sh` for the nth time (e.g. because of the
error in the first run) causes error from `rm` command: it is required
to manually restore the missing files just to remove them again.