Merged in upgrade-backend-dependencies (pull request #329)
Update `services/backend` dependencies

It includes:

* Update dependencies in services/backend, including major dependencies like Django and dj-stripe; services/workers will be provided in the separate PR)

* Adjust Python version specifier

Using >=3.8 causes the PDM to use any newer MINOR version that is available on the system, which not always can be desired. For example, on my localhost, PDM tries to use Python 3.10, while the saas-boilerplate is at the moment tested and expected to run on 3.8.*.

As such, it's recommended to disallow PDM from using any version higher than 3.8.* until we actually migrate to using more recent MINOR version for saas-boilerplate.

* Use pdm sync instead of pdm install

The latter updates the lock, which is not something desirable during the build & deploy.

* Freeze djangorestframework to 3.13.*

As defined in the https://www.django-rest-framework.org/community/release-notes/#versioning, MINOR version may include API changes (and it actually did in the 3.13.0), so it's not safe to allow for the arbitrary MINOR update (which was the case until now).

* Adjust version specifiers (many existing dependencies were allowed to update to any recent available version).

* Disable E2E tests until they are fixed (it blocked the full deployment).

After updating each package, I’ve run unit tests, and then I’ve verified the app is able to deploy successfully on the AWS. I’ve also consulted the package’s changelogs for the potential breaking changes.


Approved-by: Krzysztof Stronka